feat: add HPM5E family support
