+++ build.rs
@@ -54,6 +54,37 @@ fn get_ip_features(chip_family: &str) -> &[&str] {
             "I2C_TRANSFER_COUNT_MAX_4096",
             "ADC_BUSMODE_ENABLE_CTRL_SUPPORT",
         ],
+        "hpm5e" => &[
+            "GPTMR_MONITOR",
+            "GPTMR_OP_MODE",
+            "GPTMR_CNT_MODE",
+            "UART_RX_IDLE_DETECT",
+            "UART_FCRR",
+            "UART_RX_EN",
+            "UART_E00018_FIX",
+            "UART_9BIT_MODE",
+            "UART_ADDR_MATCH",
+            "UART_TRIG_MODE",
+            "UART_FINE_FIFO_THRLD",
+            "UART_IIR2",
+            "I2C_SUPPORT_RESET",
+            "SPI_NEW_TRANS_COUNT",
+            "SPI_CS_SELECT",
+            "SPI_SUPPORT_DIRECTIO",
+            "DMAV2_BURST_IN_FIXED_TRANS",
+            "DMAV2_BYTE_ORDER_SWAP",
+            "ADC16_HAS_MOT_EN",
+            "QEIV2_ONESHOT_MODE",
+            "QEIV2_SW_RESTART_TRG",
+            "QEIV2_TIMESTAMP",
+            "QEIV2_ADC_THRESHOLD",
+            "ENET_HAS_MII_MODE",
+            // custom
+            "DMA_V2",
+            "DMA_V2_SWAP_TABLE",
+            "I2C_TRANSFER_COUNT_MAX_4096",
+            "ADC_BUSMODE_ENABLE_CTRL_SUPPORT",
+        ],
         "hpm68" => &[
             "UART_RX_IDLE_DETECT",
             "UART_FCRR",
